First project
I've decided to recreate the classic 90's toy Simon, in your browser!
The game taught me a method of practicing repetition for the sake of
memory, so I thought this would be a fun idea.
Technologies used: HTML, CSS, JQuery
How the game works is that Simon will glow with a color, and then it's
your turn to click on the same color. Simon will then repeat that color,
along with another random color, and it's up to you to continue
remembering the long pattern of colors every turn, until you make an
error. There is no true winner, only a score for every turn you've
succeeded.
What makes this game special to me, is how I went about writing the
code. I took the time to plan out how I would execute creating the game,
and continued to break down every aspect of the game into the smallest,
simplest possible function. That way, working on it kept it's full
clarity because all of the function names describe what the functions
do. Once I created the 2 main functions (cpuGoes, playerGoes), each line
of code begins to read like English. Even better, the functions call
each other at the end of the turns, so it's ultimately a giant loop,
with the player's error being the break point.